Fold LogCasting into DefaultCasting
Given that the original authors were open to logging location in Casting, it makes sense to reduce code duplication by adding the logging code directly to DefaultCasting.

Logging is disabled by default and enabled with a boolean flag to the constructor.

A new public method isErrorLoggingEnabled() is added, which indicates whether or not logging was enabled on creation.

A new public method getErrors() is also added that returns details of any exception thrown during a cast, including the sheet, row, and column. This method throw an IllegalStateException when called on an instance without logging enabled. This ensures that a returned empty getErrors() List when logging is disabled is not misinterpreted to mean there were not casting exceptions thrown during unmarshaling.

Existing tests for DefaultCasting now include a check to ensure that logging is disabled and that the expected IllegalStateException is thrown when attempting to call getErrors(). Another set of tests that interrogates the logging will be in a forthcoming commit.
